[RF] Reduce public headers of `roobatchcompute`
After the refactoring in the last commit, the RooBatchCompute library
needs less public headers.
