+++ roofit/roofitcore/src/RooNLLVarNew.h
@@ -21,8 +21,6 @@
 
 #include <Math/Util.h>
 
-#include "RooBatchComputeTypes.h"
-
 namespace ROOT {
 namespace Experimental {
