+++ roofit/roofitmore/src/BracketAdapters.h
